pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pillow>=11.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.39.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "transformers[flax]>=4.52.3",
]
